orchestration/requirements.txt
dagster==1.5.9
dagster-webserver==1.5.9
dagster-dbt==0.21.9

# Required for DBT and Postgres integration
dbt-core==1.5.5
dbt-postgres==1.5.5
psycopg2-binary

# Optional but recommended for your pipeline
pandas
numpy
sqlalchemy
requests
kafka-python
